Extracted functionality from LwjglRenderingProcess and into new class GraphicState. Modified WorldRendererImpl to take advantage of GraphicState and initialize it appropriately. Modified LwjglRenderingProcess to refresh one of GraphicState's internal variables. Added a couple of checkstyle fixes.
